Adding tools for programming the EEPROM on the Opsis.
